developing seat reservation

+++ src/main/java/com/bamdule/studycafe/scheduler/SchedulerTask.java
@@ -1,5 +1,7 @@
 package com.bamdule.studycafe.scheduler;
 
+import com.bamdule.studycafe.config.PropertyConfig;
+import com.bamdule.studycafe.entity.reservation.service.ReservationService;
 import com.bamdule.studycafe.entity.seatusage.SeatUsageVO;
 import com.bamdule.studycafe.entity.seatusage.service.SeatUsageService;
 import com.bamdule.studycafe.websocket.MessageType;
@@ -24,16 +26,26 @@ public class SchedulerTask {
     @Autowired
     private WebSocketHandler webSocketHandler;
 
+    @Autowired
+    private PropertyConfig propertyConfig;
+
+    @Autowired
+    private ReservationService reservationService;
+
     //1분 마다 실행
     @Scheduled(cron = "0 * * * * *")
     public void checkExpiredSeatSchedule() {
 
-        LocalDateTime now = LocalDateTime.now();
-        logger.info("[MYTEST] test batch {}", now);
-        List<SeatUsageVO> expiredSeatUsages = seatUsageService.getExpiredSeatUsages(now);
-        if (expiredSeatUsages.size() > 0) {
-            seatUsageService.deleteExpiredSeatUsages(now);
-            webSocketHandler.broadcast(MessageType.EXPIRED_SEAT, expiredSeatUsages);
+        if (propertyConfig.isUseSeatSchedule()) {
+            LocalDateTime now = LocalDateTime.now();
+            logger.info("[MYTEST] test batch {}", now);
+            List<SeatUsageVO> expiredSeatUsages = seatUsageService.getExpiredSeatUsages(now);
+            if (expiredSeatUsages.size() > 0) {
+                seatUsageService.deleteExpiredSeatUsages(now);
+                webSocketHandler.broadcast(MessageType.EXPIRED_SEAT, expiredSeatUsages);
+            }
+
+            reservationService.executeReservation();
         }
 
     }
